backend/requirements.txt
# Core RAG and AI libraries
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-pinecone>=0.0.3
langchain-google-genai>=0.0.11
langchain-huggingface>=0.0.3

# Vector database and embeddings
pinecone>=3.0.0
sentence-transformers>=2.2.0

# Web frameworks
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.28.0

# Data processing
pandas>=2.0.0
pypdf>=3.17.0
python-docx>=0.8.11
openpyxl>=3.1.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
pathlib>=1.0.1

# Multi-modal processing
pillow>=10.0.0
opencv-python>=4.8.0
pytesseract>=0.3.10
pymupdf>=1.23.0
tabula-py>=2.8.0
camelot-py[cv]>=0.11.0

# Development and debugging
rich>=13.0.0

#pip install PyMuPDF opencv-python pillow pytesseract tabula-py "camelot-py[cv]"